refactor(commands): extract session-control commands into session.rs
